Added Security

+++ products/views.py
@@ -3,6 +3,7 @@ from django.contrib import messages
 from django.db.models import Q
 from .models import Product, Category
 from .forms import ProductForm
+from django.contrib.auth.decorators import login_required
 
 # Create your views here.
 
@@ -54,6 +55,7 @@ def all_products(request):
     }
     return render(request, "products/products.html", context)
 
+
 def detail(request, product_id):
     """Show Products Details"""
     product = get_object_or_404(Product, pk=product_id)
@@ -65,8 +67,13 @@ def detail(request, product_id):
     return render(request, "products/detail.html", context)
 
 
+@login_required
 def add_product(request):
     """ To add product to website """
+    if not request.user.is_superuser:
+        messages.error(request, "GET OUT!")
+        return redirect(reverse('home'))
+
     if request.method == "POST":
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -86,8 +93,13 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """ Edit Product """
+    if not request.user.is_superuser:
+        messages.error(request, "GET OUT!")
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -109,7 +121,13 @@ def edit_product(request, product_id):
         return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
+    """ Delete Product """
+    if not request.user.is_superuser:
+        messages.error(request, "GET OUT!")
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Cake removed')
